more bullet fixes and alot of ints became doubles

+++ trunk/Open GL Basic Engine/source/glutFramework/glutFramework/RVB_Entity.h
@@ -49,10 +49,10 @@ public:
 	void drawEntityTarget(int tileWidth, double scaleFactor, double mapOffsetX, double mapOffsetY);
 	
 private:
-	int xPos;
-	int yPos;
-	int targetX;
-	int targetY;
+	double xPos;
+	double yPos;
+	double targetX;
+	double targetY;
 	int health;
 	entityDirection facingDirection;
